Add Completed Assignments
When implemented this commit will add the following completed assignments
    - Clean Phone Numbers
    - Time Targeting
    - Day of the Week Targeting

Clean phone numbers assignments add the #clean_phone_numbers method. This method will take a phone number in the form of a string from a CSV file and return a formatted phone number if the phone number is a valid US phone number. This function will return "bad number" if the number is determined to be invalid.

Time target will take all the registration dates and times form the CSV file and return the most common hour of registration in a 24hr format.

Day of the Week Targeting will take all registration dates and times from the CSV file and return the most common day of the week that people registered on.